将项目的创建人设置为organizer，不再以成员的形式显示 显示个人页面的时候将负责和参与的项目同时添加进来 对不同类型的项目请求进行了规范化处理，设置为几种type格式

+++ app/cn/edu/sdu/sc/spepms/system/practice/forms/ProjectForm.java
@@ -70,7 +70,7 @@ public class ProjectForm extends AuditableModel {
     /**
      * 人数上限
      */
-    private long number;
+    private Integer number;
 
     public String getBackground() {
         return background;
@@ -136,11 +136,11 @@ public class ProjectForm extends AuditableModel {
         this.name = name;
     }
 
-    public long getNumber() {
+    public Integer getNumber() {
         return number;
     }
 
-    public void setNumber(long number) {
+    public void setNumber(Integer number) {
         this.number = number;
     }
